requirements.txt
Flask
flask-cors
camelot-py[cv]
PyPDF2
openpyxl
gunicorn
